Fix issue with populator on Laravel 4.1-dev
When the populator is registered in the app with the singleton method, it does not populate correctly the fields of a form.
I don't know why but even if it should be a singleton, each call to $app['former.populator'] would return a new empty populator. The problem disappears when using the bindShared method.

I believe it will also happens on the develop branch as the service provider is equivalent.